+++ es/index.js
@@ -1,6 +1,6 @@
 /**
-/* © 2022 Cambridge University
-/* SPDX-FileCopyrightText: 2022 Cambridge University
+/* © 2023 Cambridge University
+/* SPDX-FileCopyrightText: 2023 Cambridge University
 /* SPDX-License-Identifier: GPL-3.0-or-later
 **/
 
@@ -9,7 +9,7 @@ import '../css/pedigreejs.css';
 
 export * as pedigreejs from './pedigree.js';
 export * as pedigreejs_utils from './utils.js';
-export * as pedigreejs_form from './pedigree_form.js';
+export * as pedigreejs_form from './popup_form.js';
 export * as pedigreejs_pedcache from './pedcache.js'; 
 export * as pedigreejs_io from './io.js';
 export * as pedigreejs_canrisk_file from './canrisk_file.js';
