+++ es/widgets.js
@@ -1,13 +1,13 @@
 /**
-/* © 2022 Cambridge University
-/* SPDX-FileCopyrightText: 2022 Cambridge University
+/* © 2023 Cambridge University
+/* SPDX-FileCopyrightText: 2023 Cambridge University
 /* SPDX-License-Identifier: GPL-3.0-or-later
 **/
 
 // pedigree widgets
 import {addsibling, addchild, addparents, addpartner, rebuild, delete_node_dataset} from './pedigree.js';
 import {copy_dataset, makeid, getIdxByName} from './utils.js';
-import {save} from './pedigree_form.js';
+import {save} from './popup_form.js';
 import {current as pedcache_current} from './pedcache.js';
 
 let dragging;
